requirements.txt
absl-py==1.4.0
adamp==0.3.0
addict==2.4.0
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
basicsr==1.3.4.9
brotlipy==0.7.0
cachetools==5.3.1
certifi @ file:///croot/certifi_1690232220950/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
cftime==1.6.4
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.0
crcmod==1.7
cryptography @ file:///croot/cryptography_1689373676338/work
cycler==0.11.0
einops==0.6.1
filelock==3.12.4
fonttools==4.42.1
fsspec==2023.9.2
ftfy==6.1.1
future==0.18.3
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.57.0
huggingface-hub==0.17.3
idna @ file:///croot/idna_1666125576474/work
imageio==2.31.4
imgaug==0.4.0
importlib-metadata==6.8.0
importlib-resources==6.0.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.4.2
kiwisolver==1.4.5
kornia==0.7.0
lazy_loader==0.3
lmdb==1.4.1
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mdurl==0.1.2
MetPy==1.5.1
mkl-fft==1.3.6
mkl-random @ file:///work/mkl/mkl_random_1682950433854/work
mkl-service==2.4.0
mmengine==0.9.0
model-index==0.1.11
networkx==3.1
numpy @ file:///work/mkl/numpy_and_numpy_base_1682953417311/work
oauthlib==3.2.2
openai-clip==1.0.1
opencv-python==4.8.0.76
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.27
ordered-set==4.1.0
oss2==2.17.0
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
Pint==0.21.1
platformdirs==3.10.0
pooch==1.8.2
protobuf==4.24.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodome==3.19.0
Pygments==2.16.1
pyiqa==0.1.7
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
pyparsing==3.0.9
pyproj==3.5.0
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.8.2
pytorch-wavelets==1.3.0
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
regex==2023.8.8
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.4.2
rsa==4.9
safetensors==0.3.3
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
shapely==2.0.1
six==1.16.0
tabulate==0.9.0
tb-nightly==2.14.0a20230808
template==0.4.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==0.9.7
toml==0.10.2
tomli==2.0.1
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
tqdm==4.65.2
traitlets==5.14.3
typing_extensions @ file:///croot/typing_extensions_1690297465030/work
tzdata==2023.3
urllib3 @ file:///croot/urllib3_1686163155763/work
wcwidth==0.2.6
Werkzeug==2.3.7
xarray==2023.1.0
yacs==0.1.8
yapf==0.40.2
zipp==3.16.2
